LayoutTests/imported/w3c/web-platform-tests/wai-aria/role/contextual-roles.tentative-expected.txt

x
x

PASS role is sectionfooter (in main)
PASS role is sectionheader (in main)
